# docs/public/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PricingResponse": {
        "description": "Response model for model pricing.",
        "properties": {
          "created_at": {
            "title": "Created At",
            "type": "string"
          },
          "effective_at": {
            "title": "Effective At",
            "type": "string"
          },
          "input_price_per_million": {
            "title": "Input Price Per Million",
            "type": "number"
          },
          "model_key": {
            "title": "Model Key",
            "type": "string"
          },
          "output_price_per_million": {
            "title": "Output Price Per Million",
            "type": "number"
          },
          "updated_at": {
            "title": "Updated At",
            "type": "string"
          }
        },
        "required": [
          "model_key",
          "effective_at",
          "input_price_per_million",
          "output_price_per_million",
          "created_at",
          "updated_at"
        ],
        "title": "PricingResponse",
        "type": "object"
      },
...
      "SetPricingRequest": {
        "description": "Request model for setting model pricing.",
        "properties": {
          "effective_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "ISO 8601 datetime from which this price applies. Defaults to now if omitted.",
            "title": "Effective At"
          },
          "input_price_per_million": {
            "description": "Price per 1M input tokens",
            "minimum": 0.0,
            "title": "Input Price Per Million",
            "type": "number"
          },
          "model_key": {
            "description": "Model identifier in format 'provider:model'",
            "title": "Model Key",
            "type": "string"
          },
          "output_price_per_million": {
            "description": "Price per 1M output tokens",
            "minimum": 0.0,
            "title": "Output Price Per Million",
            "type": "number"
          }
        },
        "required": [
          "model_key",
          "input_price_per_million",
          "output_price_per_million"
        ],
        "title": "SetPricingRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "A clean FastAPI gateway for any-llm with API key management",
    "title": "any-llm-gateway",
    "version": "0.0.0-dev"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/pricing": {
      "get": {
        "description": "List all model pricing.",
        "operationId": "list_pricing_v1_pricing_get",
        "parameters": [
          {
            "in": "query",
            "name": "skip",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Skip",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PricingResponse"
                  },
                  "title": "Response List Pricing V1 Pricing Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Pricing",
        "tags": [
          "pricing"
        ]
      },
      "post": {
        "description": "Set or update pricing for a model.",
        "operationId": "set_pricing_v1_pricing_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetPricingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Set Pricing",
        "tags": [
          "pricing"
        ]
      }
    },
    "/v1/pricing/{model_key}": {
      "delete": {
        "description": "Delete pricing entries for a model.",
        "operationId": "delete_pricing_v1_pricing__model_key__delete",
        "parameters": [
          {
            "in": "path",
            "name": "model_key",
            "required": true,
            "schema": {
              "title": "Model Key",
              "type": "string"
            }
          },
          {
            "description": "ISO datetime identifying a specific pricing row to delete",
            "in": "query",
            "name": "effective_at",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO datetime identifying a specific pricing row to delete",
              "title": "Effective At"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Pricing",
        "tags": [
          "pricing"
        ]
      },
      "get": {
        "description": "Get pricing for a specific model as of a timestamp.",
        "operationId": "get_pricing_v1_pricing__model_key__get",
        "parameters": [
          {
            "in": "path",
            "name": "model_key",
            "required": true,
            "schema": {
              "title": "Model Key",
              "type": "string"
            }
          },
          {
            "description": "ISO datetime for effective lookup",
            "in": "query",
            "name": "as_of",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "ISO datetime for effective lookup",
              "title": "as_of"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PricingResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pricing",
        "tags": [
          "pricing"
        ]
      }
    },
    "/v1/pricing/{model_key}/history": {
      "get": {
        "description": "Return the full pricing history for a model.",
        "operationId": "get_pricing_history_v1_pricing__model_key__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "model_key",
            "required": true,
            "schema": {
              "title": "Model Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PricingResponse"
                  },
                  "title": "Response Get Pricing History V1 Pricing  Model Key  History Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Pricing History",
        "tags": [
          "pricing"
        ]
      }
    },
...
  }
}